requirements.txt
# newcomer-learning

streamlit==1.37.1
openai==1.93.3
streamlit_autorefresh
faiss-cpu
sentence-transformers
